ui: Make the collection icon view more compact
Don't make the game thumbnails expand as the window expand but rather
pack as much as possible in the center of the window. This avoid a weird
accordion effect when resizing the window and avoid having to move the
eyes as much to see the available games.

Also reduce a bit the spacing of the rows and columns to make the view
more compact.