developer: Add 'has-loaded' property
This will be used in subsequent commits to get rid of changed() signal.

Use 'true' value for GenericDeveloper, because it's a simple string that
is always available.

Use 'false' for DummyDeveloper, because it doesn't have any value.

Use 'notify::has-loaded' instead of 'connect' signal in DatabaseDeveloper,
and connect signal lazily because it's impossible to handle already loaded
(such as GenericDeveloper) case otherwise.