platforms-view: Make adaptive
Add a leaflet to PlatformsView. When folded, open subview when clicking
on a platform.

Add a second headerbar to CollectionHeaderBar, show it when viewing a
platform when the window is narrow. This headerbar contains a back button
and another search button.

Implement Alt+Left (or Alt+Right, depending on text direction) shortcut
to go back to platforms list when folded.

Re-enable selection for narrow mode when using gamepad.