savestate: Store screenshot aspect ratio
retro-gtk provides screenshot aspect ratio in form of x-dpi and y-dpi
options in the pixbuf. We try to save them, but fail: they aren't actually
written in the file. Hence when saving a new screenshot, read its aspect
ratio right away and store it in the savestate keyfile.

This will be used in the next commit to use proper scpect ratio for
thumbnails.