retro-runner: Add functions for extra metadata
Savestate subclasses will need to load their metadata. Since currently
all metadata is provided at once via set_metadata_manual() or
set_metadata_automatic() and is written right there, any extra metadata
for subclasses needs to be supplied before it, and it will be written via
a virtual method in Savestate itself, added in the following commit.

This will be used in the following commits to store extra metadata for
Nintendo DS games.