nintendo-ds: Add NintendoDsSavestate
This special savestate stores screen layout and which screen is visible
if quick switch is used.

NintendoDsRunner is supposed to override load_extra_savestate_metadata(),
call load_extra_metadata() from it and use the property values.

Similarly, it overrides save_extra_savestate_metadata() and change the
properties, so that later they are written into the keyfile.