database: Cache games
Currently, when starting the app we load everything anew. Instead, cache
the current state in the database and load it immediately when starting
the app. Then, run the current discovery process (since it's threaded,
there are no UI stalls) and amend the game collection and the database.

This should significantly speed up game loading.

Remove pausing, as loading should be fast enough now. If it becomes a
problem, it can be restored separately.