Add search provider
Since games are now cached in the database, it's possible to implement a
search provider as a separate binary that will only peek into the database
and cover cache.

Searching by multiple search terms is a bit tricky in SQL as the query
changes based on the number of search terms. Because of that, the query has
to be generated on the fly before binding values to it. Additionally, the
`ORDER BY` sorting doesn't completely match Games, so it can't be used.
Instead, games are just retrieved unsorted and then sorted manually with
`collate()`. This also means thatthe search provider has to have the
correct locale set, so we have to link it with GTK and use Gtk.Application
instead of GLib.Application.

Stop cleaning up dbus directory for flatpak, as the search provider service
is installed there.

Fixes https://gitlab.gnome.org/GNOME/gnome-games/issues/45