display-view: Recreate the runner when restarting
Currently restart just asks the core to restart and it's up to the cores
to support it. Some cores, such as BlastEm or MAME, don't handle it.

To circumvent it, destroy the current runner and create a new one, that
is guaranteed to work.

Additionally, adds and shows a dialog in case the current game cannot be saved,
same as when quitting the game.