local-cover: Make the basename check more strict
Currently we take the basename of a ROM, strip its extension, and look for
files that start with the same prefix. This breaks as soon as we have names
as following:

 * Game.rom
 * Game - Something Else.jpg

Obviously they both contain the prefix "Game", so the file gets picked as
the cover, which is incorrect. Instead, take prefix of both files and check
that the whole prefix matches and not just the beginning.