games-page: Add support for selection-mode
It now stores the selected games in a GenericSet. Whenever a game's
checked state changes it adds/removes the corresponding GameIconView
into the selected_games list.

When in selection-mode activating the GameIconView will toggle the
game's checked state.

All or none games can be selected using the select_all/none ()
When in PlatformsPage all games of the currently viewed platform
can be selected by calling select_all () in PlatformsPage.

select_none () will remove every game in selected_games regardless
of the currenlty selected platform.

This will be used in the upcoming commits to implement selection-mode
in CollectionView.