collection-view: Add selection button to headerbar
On enabling selection-mode the main headerbar will be inaccessible until
the selection is completed/cancelled.

When in PlatformsPage, if view is transitioned from desktop to mobile
view, the view will be locked to the currently selected platform. Swipe
back will be disabled until selection is completed/cancelled.

Select all in PlatformsView will only select games belonging to the
currently selected platform.

This will be used in later commits to implement CollectionsPage