collection: Introduce CollectionManager
CollectionManager will be used to handle the creation and management
of collections. It consists of a HashTable which stores all the
discovered collections, a FavoriteCollection which is a non-user
collection for managing favorite games.

Provides a toggle_favorite() which adds and removes a game depending
on the list of games