games-page: Explicitly clear selected_games on select_none()
This is useful when games are removed from a collection being viewed.
When a game is removed from a collection, its GameIconView's checked
state cannot be updated because the game is removed from the GameModel
and the corresponding GameIconView is no longer available for setting
the checked prop to false. And so the game remains in the selected_games
set even after removal. We can call remove_all() on selected_games to
ensure it will always clear the selected_games set.