game: Add last_played property
last_played is a DateTime which is used to store when a game was last
played. If this prop is null then it implies the game was never played
after the introduction of this prop.

Also adds a static compare_by_date_time which can be used to sort the
games in decreasing order of last played.