game-model: Add support for different sort types
Adds a SortType enum which can be used to decide which comparison func
to use when comparing games in the game-model. It automatically calls
sort when sorting type is changed.

This will be used in the upcoming commits to sort the games in the
decreasing order of last played games.