collection-thumbnail: Refresh the covers whenever collection changes
When a collection-model's sort changes, the flowbox it is bound to has
to refresh to conform to changes in the collection-model. When this
happens thw flowbox will change the collection that belongs to
collection-icon-view which then changes the collection prop here in
collection-thumbnail. But the thumbnail wasn't refreshed to accomodate
the change in collection-model of the new collection. This commit makes
sure that thumbnails are refreshed whenever the collection prop changes.