collection-view: Check subpage is open before consuming input
Currently when an empty collection is opened the is_collection_empty
property is set to false. This property alone was used to check if
search should be disabled. That means whenever an empty collection was
viewed and exited, search would no longer work because is_collection_
empty prop is still true. So we now check whether collection subpage
is open before checking if the current collection is empty to consume
the input, so that search is disabled in those situations.

This bug can't be seen as of now because we currently only support
non-user collections which are hidden when empty.